[configuration] set much lower leader weight for non-BR devices
There can be cases that both a phone and accessory can be a leader
of their own partitions and then merges together when the user
moves close to the accessory. In this case, it's better for the
phone to always attach to the accessory as a Child rather than
being a Leader.

To achieve this, we can set a leader weight less than the default
value 64 - currently, 32 for non-BR Android devices.

Bug: 404979710
Change-Id: I63bff20e9945751449299545a03f52764e28f9b8